:stethoscope: Heal Unit Logic (#33)
* :sparkles: Implemeneted attack logic for heal unit

* :rocket: Adjusted attack function for EarthTank & EarthSoldier

* :rocket: Adjusted attack function for EarthSoldier

* :sparkles: Changed attack logic for unit classes

* :speech_balloon: Added printing for maintenance list

* :green_heart: Fixed bug for picking healing unit

* :zap: Changed health to double

* :hammer: Fixed missing line in HealUnit

* :boom: Added line to nullify pointer causing a bug with enemy selection

* :boom: Added line to nullify pointer causing a bug with enemy selection

* :memo: Added setter for UML Join Time

* :bulb: Added comments for unit classes

* :bug: added a missing bracket

* :hammer: Fixed heal priority

---------

+++ UnitClasses/AlienDrone.h
@@ -7,9 +7,10 @@
 class AlienDrone: public Unit
 {
 public:
-	AlienDrone(Game*, int, int, int);
-	void printFought();
-	void attack();
+	AlienDrone(Game*, double, int, int);
+
+	void printFought(); // Print the units that the drone has fought
+	void attack(); // Attack the enemy units
 };
 
 #endif
